move auth to a controller
